Fix RESIT edge elimination
In the current version of RESIT the independence of the residuals is not tested properly: the independence has to be tested between the residuals and ALL the potential parents, including 'l' (predictors + [l], line 173).

Also, when there is no potential parent except 'l' the independence between Xk and 'l' must be tested in order to exclude 'l' as a parent of Xk. Otherwise it would not be possible to recover causal graphs that are not connected.